[Refactor] Move sidebar items into folders
